Rename cmd/apiserver -> cmd/maya-apiserver
