+++ cmd/maya-apiserver/app/mockit/sample_hello_world.yaml
